Refactor content to new homeschooltracker.content package
