+++ src/test/java/homeschooltracker/content/LessonTest.java
@@ -1,4 +1,4 @@
-package homeschooltracker;
+package homeschooltracker.content;
 
 import junit.framework.TestCase;
